[RegistryPreview] Fix showing in the background on start
