+++ src/modules/Hosts/Hosts/HostsXAML/MainWindow.xaml.cs
@@ -21,7 +21,8 @@ namespace Hosts
             AppWindow.SetIcon("Assets/Hosts/Hosts.ico");
             Title = ResourceLoaderInstance.ResourceLoader.GetString("WindowTitle");
 
-            BringToForeground();
+            var handle = this.GetWindowHandle();
+            ManagedCommon.WindowHelpers.BringToForeground(handle);
 
             Activated += MainWindow_Activated;
         }
@@ -35,26 +36,6 @@ namespace Hosts
             else
             {
                 AppTitleTextBlock.Foreground = (SolidColorBrush)App.Current.Resources["WindowCaptionForeground"];
-            }
-        }
-
-        private void BringToForeground()
-        {
-            var handle = this.GetWindowHandle();
-            var fgHandle = NativeMethods.GetForegroundWindow();
-
-            var threadId1 = NativeMethods.GetWindowThreadProcessId(handle, System.IntPtr.Zero);
-            var threadId2 = NativeMethods.GetWindowThreadProcessId(fgHandle, System.IntPtr.Zero);
-
-            if (threadId1 != threadId2)
-            {
-                NativeMethods.AttachThreadInput(threadId1, threadId2, true);
-                NativeMethods.SetForegroundWindow(handle);
-                NativeMethods.AttachThreadInput(threadId1, threadId2, false);
-            }
-            else
-            {
-                NativeMethods.SetForegroundWindow(handle);
             }
         }
     }
